fix header component and small semantic changes in children components
